[WIP] ISS74: Training pytests (#76)
* added PyTorch and Keras DataGenerators

* adding pyyaml to reqts

* fixing keras imports

* fixing model_io import

* completed solaris.nets.train.Trainer.__init__()

* completed solaris.nets.train.Trainer.__init__()

* resolving conflict with dev

* stash

* adding in ModelCheckpoint callback for torch

* intermediate - transfer

* preliminary code completed for model training

* adding metrics pytests

* fixing rio-tiler dependency

* fixing rio-tiler dependency in environment.yml

* adding tests for get_callbacks()

* pytests for keras LR scheduler

* adding tests for torch lr scheduler

* adding pytests for re-implemented transforms

* adding test for build__pipeline()

* adding test for get_optimizer()

* adding pytests for solaris.nets.losses

* debugging alongside test_losses

* adding pytests for keras datagen

* finished torch pytest for datagen

+++ solaris/nets/metrics.py
@@ -42,7 +42,7 @@ def precision(y_true, y_pred):
     truth = K.round(K.clip(y_true, K.epsilon(), 1))
     pred_pos = K.round(K.clip(y_pred, K.epsilon(), 1))
     true_pos = K.sum(K.cast(K.all(K.stack([truth, pred_pos], axis=2), axis=2),
-                            dtype='float32'))
+                            dtype='float64'))
     pred_pos_ct = K.sum(pred_pos) + K.epsilon()
     precision = true_pos/pred_pos_ct
 
@@ -59,7 +59,7 @@ def recall(y_true, y_pred):
     truth = K.round(K.clip(y_true, K.epsilon(), 1))
     pred_pos = K.round(K.clip(y_pred, K.epsilon(), 1))
     true_pos = K.sum(K.cast(K.all(K.stack([truth, pred_pos], axis=2), axis=2),
-                            dtype='float32'))
+                            dtype='float64'))
     truth_ct = K.sum(K.round(K.clip(y_true, K.epsilon(), 1)))
     if truth_ct == 0:
         return 0
